Create Switch
Complete the getLetter(s) function in the editor. It has one parameter: a string, , consisting of lowercase English alphabetic letters (i.e., a through z). It must return A, B, C, or D depending on the following criteria:

If the first character in string  is in the set a,e,i,o,u, then return A.
If the first character in string  is in the set b,c,d,ef,g , then return B.
If the first character in string  is in the set h,j,k,l,m, then return C.
If the first character in string  is in the set n,p,q,r,s,t,v,w,x,y,z, then return D.